Fire input/change events from type() and clear().
The old approach of using textInput events worked for type() but not for
clear(). It looks like textInput events with an empty data field don't
publish the expected browser events, so anything listening for changes
to the element's value (e.g. Angular) wouldn't see anything.

Firing "input" and "change" should give the expected behavior: "input"
to trigger on-keypress handlers, and "change" to trigger on-enter
handlers.