Upgrade to analyzer 0.33.3
Switch to use a temporary `ParsedLibraryResult` API until the final API
is worked out in `package:build`. The current approach of using
`computeNode()` will break in the next breaking release of `build` when
it flips to using `AnalysisDriver` isntead of `AnalysisContext` to back
the `Resolver`. This temporary API will allow backwards compatibility
until the migration is complete.